Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240403/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.c
@@ -3,7 +3,7 @@
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    printf("%d\n%.0f", a+b+c, (float)(a+b+c)/3);
+    printf("%d\n%d", a+b+c,(a+b+c)/3);
     // 여기에 코드를 작성해주세요.
     return 0;
 }
